% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_age_sex_plots.R
\name{compute_de_cor_mat}
\alias{compute_de_cor_mat}
\title{Compute correlation matrix across cell_type x region groups}
\usage{
compute_de_cor_mat(
  de_dt,
  cell_types_use,
  regions_use,
  non_neuron_types,
  fdr_cutoff = 0.05
)
}
\arguments{
\item{de_dt}{A prepared DE data.table from prep_de/read_de_results.}

\item{cell_types_use}{Character vector of cell types to include.}

\item{regions_use}{Character vector of regions to include.}

\item{non_neuron_types}{Character vector of non-neuronal cell types.}

\item{fdr_cutoff}{Adjusted p-value threshold.}
}
\value{
A square numeric matrix with dimnames "cell_type__region".
}
\description{
Correlations are computed on log_fc among genes passing an FDR filter in either
of the compared groups. The returned value is signed rho^2.
}
