% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_age_sex_plots.R
\name{extract_donor_ages}
\alias{extract_donor_ages}
\title{Extract donor ages from cell metadata}
\usage{
extract_donor_ages(cell_metadata)
}
\arguments{
\item{cell_metadata}{A data.table containing donor_external_id and age.}
}
\value{
A named numeric vector of donor ages.
}
\description{
Extract donor ages from cell metadata
}
