% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_age_sex_plots.R
\name{plot_donor_gex_age_scatterplot}
\alias{plot_donor_gex_age_scatterplot}
\title{Donor-level GEX scatterplot vs age
(originally plot_donor_gex_age_scatterplot)}
\usage{
plot_donor_gex_age_scatterplot(exp_vector, donor_ages, main = "")
}
\arguments{
\item{exp_vector}{Named numeric vector of expression values (names are donor IDs).}

\item{donor_ages}{Named numeric vector of donor ages (names are donor IDs).}

\item{main}{Character plot title.}
}
\value{
Invisibly returns NULL.
}
\description{
Donor-level GEX scatterplot vs age
(originally plot_donor_gex_age_scatterplot)
}
