% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_age_sex_plots.R
\name{prep_de}
\alias{prep_de}
\title{Prepare DE results for plotting and downstream analysis
(originally format_de_results)}
\usage{
prep_de(df, gene_to_chr)
}
\arguments{
\item{df}{A data.frame-like object returned by parse_de_inputs.}

\item{gene_to_chr}{A data.table with at least "gene" and "chr".}
}
\value{
A data.table with standardized columns and annotations.
}
\description{
Prepare DE results for plotting and downstream analysis
(originally format_de_results)
}
