% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_age_sex_plots.R
\name{prep_region_lfc_matrix}
\alias{prep_region_lfc_matrix}
\title{Prepare region-specific LFC matrix
(originally prepare_region_lfc_matrix)}
\usage{
prep_region_lfc_matrix(de_dt, genes_use, cell_types_use, regions_use)
}
\arguments{
\item{de_dt}{Prepared DE data.table.}

\item{genes_use}{Genes to retain.}

\item{cell_types_use}{Character vector of cell types.}

\item{regions_use}{Character vector of regions.}
}
\value{
A numeric matrix of log fold-changes.
}
\description{
Prepare region-specific LFC matrix
(originally prepare_region_lfc_matrix)
}
