% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_age_sex_plots.R
\name{read_cell_metadata}
\alias{read_cell_metadata}
\title{Read cell metadata}
\usage{
read_cell_metadata(cell_metadata_file)
}
\arguments{
\item{cell_metadata_file}{Path to annotated cell metadata file.}
}
\value{
A data.table of cell metadata.
}
\description{
Read cell metadata
}
