% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_age_sex_plots.R
\name{read_gene_to_chr}
\alias{read_gene_to_chr}
\title{Read gene-to-chromosome mapping}
\usage{
read_gene_to_chr(gene_to_chr_file)
}
\arguments{
\item{gene_to_chr_file}{Path to a tabular file readable by data.table::fread.}
}
\value{
A data.table with columns including gene and chr.
}
\description{
The input file must contain at least columns "gene" and "chr".
Chromosomes are filtered to 1:22, X, Y, and M by default.
}
