% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_age_sex_plots.R
\name{read_metacells}
\alias{read_metacells}
\title{Read and aggregate metacells
(originally load_metacells)}
\usage{
read_metacells(
  path,
  cell_types_use,
  regions_use = c("CaH", "Pu", "NAC", "ic", "DFC")
)
}
\arguments{
\item{path}{Path to metacell count matrix file.}

\item{cell_types_use}{Character vector of cell types to retain.}

\item{regions_use}{Character vector of regions to retain.}
}
\value{
A list with elements:
  - metacells: dgCMatrix of aggregated TPM
  - col_metadata: data.table of donor/cell_type/region
}
\description{
Read and aggregate metacells
(originally load_metacells)
}
