% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_age_sex_plots.R
\name{split_metacells_by_cell_type_region}
\alias{split_metacells_by_cell_type_region}
\title{Split metacells by cell_type and region}
\usage{
split_metacells_by_cell_type_region(metacells, col_metadata, donor_ages)
}
\arguments{
\item{metacells}{dgCMatrix of TPM values.}

\item{col_metadata}{data.table returned by read_metacells.}

\item{donor_ages}{Named numeric vector of donor ages.}
}
\value{
A named list of dgCMatrix objects, sorted by donor age.
}
\description{
Split metacells by cell_type and region
}
