% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_age_sex_plots.R
\name{summarize_metacells}
\alias{summarize_metacells}
\title{Summarize metacells by gene and cell_type-region}
\usage{
summarize_metacells(metacells, col_metadata, donor_ages)
}
\arguments{
\item{metacells}{dgCMatrix of TPM values.}

\item{col_metadata}{data.table returned by read_metacells.}

\item{donor_ages}{Named numeric vector of donor ages.}
}
\value{
A data.table summary.
}
\description{
Summarize metacells by gene and cell_type-region
}
