% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bican.mccarroll.de.analysis-package.R
\docType{package}
\name{bican.mccarroll.de.analysis-package}
\alias{bican.mccarroll.de.analysis}
\alias{bican.mccarroll.de.analysis-package}
\title{bican.mccarroll.steveb}
\description{
Steve Burger will add text here.
}
\author{
\strong{Maintainer}: Steven Burger \email{sburger@broadinstitute.org}

Authors:
\itemize{
  \item James Nemesh \email{nemesh@broadinstitute.org}
}

}
\keyword{internal}
